Create new emailblasts targeting module
